requirements.txt
pymongo[srv]==4.2.0
pandas 
numpy
matplotlib
seaborn
scikit-learn
scikit-image
Flask
#-e .
certifi
python-dotenv
from-root
PyYAML